refactor(storage): SQLite storageクラスをファイル分割
- SqliteWorkoutSessionStorage.ts (321行) に分離
- SqliteExerciseLibraryStorage.ts (199行) に分離
- storage.tsをインデックスファイル (7行) に変更
- updateLastUsedメソッド名を契約に合わせて修正
- DatabaseManager型エラーを修正

Before: storage.ts 521行 → After: 3ファイルで保守性向上
